close #1; update menu; responsibility of rreservation dialog
