+++ src/app/componentes/dialogs/dialog-disponibilidade/dialog-disponibilidade.component.html
@@ -4,9 +4,9 @@
 
    <mat-step [stepControl]="formTermsGroup" [completed]="false" label="Termos">
       <form [formGroup]="formTermsGroup" class="flex flex-col">
-        <p class="py-2">O Preenchimento do formulário a seguir não é garantia de que a reserva foi efetuada. A Pousada ao receber as informações e verificando a disponibilidade conforme sua demanda entrará em contato para que se possa dar continuidade a reserva.</p>
+        <p class="py-2">O Preenchimento do formulário a seguir não é garantia de que a reserva será efetuada. A Pousada ao receber as informações desse formulário e verificando a disponibilidade conforme sua demanda entrará em contato para dar continuidade ao processo de reserva.</p>
 
-        <p class="py-2">Dessa forma solicitamos que o formulário seja preenchido as informações corretas para que possamos entrar em contato tão logo quanto seja possível.</p>
+        <p class="py-2">Dessa forma solicitamos que o formulário seja preenchido com as informações corretas de forma que possamos entrar em contato tão logo quanto seja possível.</p>
 
         <label class="py-4 inline-flex items-center">
           <input
