+++ src/app/componentes/header/header.component.html
@@ -29,6 +29,7 @@
         <li><a href="/#cidade-rioclaro-rj">Rio Claro</a></li>
         <li><a href="/#novidades">Notícias</a></li>
         <li><a href="/#fale-com-a-pousada">Fale Conosco</a></li>
+        <li><a (click)="openReservationDialog()" >Reservas</a></li>
       </ul>
     </nav>
   </div>
