+++ src/app/componentes/header/header.component.scss
@@ -1,0 +1,17 @@
+/* # Main Menu
+================================================== */
+.main-menu{
+  @apply w-full flex flex-col sm:flex-row items-center justify-end font-bold md:font-semibold uppercase;
+
+  li{
+    a{
+      @apply block mx-1 py-2 px-3 rounded-md cursor-pointer transition-all duration-500 ease-in-out hover:bg-heavy-metal-200;
+    }
+
+    &:last-child{
+      a{
+        @apply mr-0 bg-heavy-metal-700 text-white border hover:bg-heavy-metal-500 focus:bg-heavy-metal-500 focus:outline-none hover:outline-none focus:ring-2 hover:ring-2 focus:ring-offset-2 hover:ring-offset-2 focus:ring-heavy-metal-500 hover:ring-heavy-metal-500;
+      }
+    }
+  }
+}
