+++ src/app/componentes/section-cidade/section-cidade.component.html
@@ -20,8 +20,8 @@
     </div>
   </div>
 
-  <ul class="flex flex-row flex-wrap py-6 -mx-4">
-    <li class="w-1/5 p-4">
+  <ul class="flex flex-row flex-wrap py-6 -mx-4 justify-center">
+    <li class="w-6/12 sm:w-4/12 md:w-1/5 p-4">
       <a class="default-image default-image-size">
         <ngx-picture
         src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942220/pousada-muller-site/galeria-cidade/1A5918B7CAC314AEC2EE8F515EF2ECAB03FC8D13E9A26EC6D86DE8AC88260FFF_ewlqas.WebP"
@@ -31,7 +31,7 @@
       </a>
     </li>
 
-    <li class="w-1/5 p-4">
+    <li class="w-6/12 sm:w-4/12 md:w-1/5 p-4">
       <a class="default-image default-image-size">
         <ngx-picture
         src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942216/pousada-muller-site/galeria-cidade/01B1EE4CBEB5E12D7F78CDAA1779B8F17162871DAD9E00D5224A65F481282128_tacde1.WebP"
@@ -41,7 +41,7 @@
       </a>
     </li>
 
-    <li class="w-2/5 p-4">
+    <li class="w-6/12 sm:w-4/12 md:w-1/5 lg:w-2/5 p-4">
       <a class="default-image default-image-size">
         <ngx-picture
         src="https://res.cloudinary.com/miguelsmuller/image/upload/ar_2:1,c_crop,w_1890/v1621942205/pousada-muller-site/galeria-cidade/E2711845540D7FD337CC294E3074297DCF792CC4D4BFF85FE3EE1A2198F7DEC4_v8wobm.webp"
@@ -51,7 +51,7 @@
       </a>
     </li>
 
-    <li class="w-1/5 p-4">
+    <li class="w-6/12 sm:w-4/12 md:w-1/5 p-4">
       <a class="default-image default-image-size">
         <ngx-picture
         src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942206/pousada-muller-site/galeria-cidade/D8979BDBA8D139242AFD46A0C9F04B3C42FC87C419DDD8D3E53ED3C345BFDE53_qcawez.WebP"
@@ -61,7 +61,7 @@
       </a>
     </li>
 
-    <li class="w-2/5 p-4">
+    <li class="w-6/12 sm:w-4/12 md:w-1/5 lg:w-2/5 p-4">
       <a class="default-image default-image-size">
         <ngx-picture
         src="https://res.cloudinary.com/miguelsmuller/image/upload/ar_2:1,c_crop,w_1890/v1621942204/pousada-muller-site/galeria-cidade/0E6FECE54A113833F167B0DCFC83D3AE078B6EB4A13D4AC5D598B5054DE74411_mr19e1.WebP"
@@ -71,7 +71,7 @@
       </a>
     </li>
 
-    <li class="w-1/5 p-4">
+    <li class="w-6/12 sm:w-4/12 md:w-1/5 p-4">
       <a class="default-image default-image-size">
         <ngx-picture
         src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942205/pousada-muller-site/galeria-cidade/A992BD0E9D7BCF089F04D999FD86970ACEDF2DB3507FDFC2A5733B4775B25732_p01ixb.WebP"
@@ -81,7 +81,7 @@
       </a>
     </li>
 
-    <li class="w-1/5 p-4">
+    <li class="w-6/12 sm:w-4/12 md:w-1/5 p-4">
       <a class="default-image default-image-size">
         <ngx-picture
         src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942210/pousada-muller-site/galeria-cidade/704BB6C438B9FB7ED46DFD756FE19D5E42BF66ACEF7093F20C061824646D0EBD_lfqcv6.WebP"
@@ -91,7 +91,7 @@
       </a>
     </li>
 
-    <li class="w-1/5 p-4">
+    <li class="w-6/12 sm:w-4/12 md:w-1/5 p-4">
       <a class="default-image default-image-size">
         <ngx-picture
         src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942211/pousada-muller-site/galeria-cidade/60B2F474A7868A58BA9B3DAEA46AE31F72E518F4A0097E4C394A5F06763A8E33_dzqckx.WebP"
