+++ src/app/componentes/section-contato/section-contato.component.ts
@@ -32,7 +32,7 @@ export class SectionContatoComponent implements OnInit {
   }
 
   sendReservation(formData: FormGroup): void{
-    this.emailService.doCommunication(formData).subscribe(
+    this.emailService.senfForm(formData).subscribe(
       susc => { this.formStatus = 'Sua manifestação foi enviada com sucesso.'; },
       erro => { console.log(erro); this.formStatus = erro.message; }
     );
