+++ src/app/componentes/section-cover/section-cover.component.html
@@ -8,7 +8,7 @@
         <p class="py-2 text-base font-normal">Aproveite a tranquilidade e beleza que a região de Rio Claro pode oferecer!</p>
       </div>
 
-      <button class="default-button my-4" (click)="openDialogDisponibilidade()">Ver Disponibilidade</button>
+      <button class="default-button my-4" (click)="openReservationDialog()">Realizar Agendamento</button>
     </div>
 
     <div class="flex-1 flex md:justify-end -mx-4 md:ml-10 md:mx-0 py-6 sm:py-0">
