+++ src/app/componentes/section-cover/section-cover.component.ts
@@ -1,35 +1,17 @@
-import { Component, OnInit } from '@angular/core';
-
-import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
-
-import { DialogDisponibilidadeComponent } from '@app/componentes/dialogs/dialog-disponibilidade/dialog-disponibilidade.component';
+import { Component } from '@angular/core';
+import { EmailService } from '@app/shared/email.service';
 
 @Component({
   selector: 'app-section-cover',
   templateUrl: './section-cover.component.html',
   styleUrls: ['./section-cover.component.scss']
 })
-export class SectionCoverComponent implements OnInit {
+export class SectionCoverComponent {
+  constructor (
+    private emailService: EmailService
+  ){}
 
-  constructor(
-    private dialog: MatDialog,
-  ) { }
-
-  ngOnInit(): void {
+  openReservationDialog() {
+    this.emailService.openReservationDialog();
   }
-
-  openDialogDisponibilidade() {
-    const formConfig = new MatDialogConfig();
-    formConfig.panelClass = 'mat-dialog';
-    formConfig.disableClose = false;
-    formConfig.autoFocus = true;
-    formConfig.data = '';
-
-    const formularioCriarCargo = this.dialog.open(DialogDisponibilidadeComponent, formConfig);
-
-    formularioCriarCargo.afterClosed().subscribe((data) => {
-      //console.log(data);
-    });
-  }
-
 }
