+++ src/app/componentes/section-historia/section-historia.component.html
@@ -16,8 +16,9 @@
     </div>
   </div>
 
-  <ul class="flex flex-row py-6 -mx-4">
-    <li class="w-1/5 p-4">
+
+  <ul class="flex flex-row flex-wrap py-6 -mx-4 justify-center">
+    <li class="w-6/12 sm:w-4/12 lg:w-3/12 xl:w-1/5 p-4">
       <a href="#" class="default-image">
         <ngx-picture
           src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942224/pousada-muller-site/galeria-pousada/pousada-1-square_tkk9dm.WebP"
@@ -26,7 +27,7 @@
         ></ngx-picture>
       </a>
     </li>
-    <li class="w-1/5 p-4">
+    <li class="w-6/12 sm:w-4/12 lg:w-3/12 xl:w-1/5 p-4">
       <a href="#" class="default-image">
         <ngx-picture
           src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942224/pousada-muller-site/galeria-pousada/pousada-2-square_c5fnyb.WebP"
@@ -35,7 +36,7 @@
         ></ngx-picture>
       </a>
     </li>
-    <li class="w-1/5 p-4">
+    <li class="w-6/12 sm:w-4/12 lg:w-3/12 xl:w-1/5 p-4">
       <a href="#" class="default-image">
         <ngx-picture
           src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942228/pousada-muller-site/galeria-pousada/pousada-5-square_blehpf.WebP"
@@ -44,7 +45,7 @@
         ></ngx-picture>
       </a>
     </li>
-    <li class="w-1/5 p-4">
+    <li class="w-6/12 sm:w-4/12 lg:w-3/12 xl:w-1/5 p-4">
       <a href="#" class="default-image">
         <ngx-picture
           src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942222/pousada-muller-site/galeria-pousada/pousada-6-square_cvbpvq.WebP"
@@ -53,7 +54,7 @@
         ></ngx-picture>
       </a>
     </li>
-    <li class="w-1/5 p-4">
+    <li class="w-6/12 sm:w-4/12 lg:w-3/12 xl:w-1/5 p-4">
       <a href="#" class="default-image">
         <ngx-picture
           src="https://res.cloudinary.com/miguelsmuller/image/upload/w_1000,ar_1:1,c_fill,g_auto/v1621942223/pousada-muller-site/galeria-pousada/pousada-7-square_twyhh9.jpg"
