+++ src/theme/_tailwind.scss
@@ -18,8 +18,7 @@
   }
 
   .default-image{
-    //p-0.5 - border-heavy-metal-200
-    @apply block box-content bg-white rounded-lg shadow-md border;
+    @apply block box-content p-0.5 border-heavy-metal-200 bg-white rounded-lg shadow-md border;
 
     .ngx-picture__picture {
       @apply max-h-full h-auto w-auto;
